fix bugs

+++ codes/data_scripts/combine_frames.py
@@ -8,7 +8,6 @@ import numpy as np
 
 sys.path.append(osp.dirname(osp.dirname(osp.abspath(__file__))))
 import utils.util as util
-import utils.yuv_io as yuv_io
 
 
 def combine_frames_to_single_video(frame_path, video_path):
@@ -29,16 +28,18 @@ if __name__ == '__main__':
         video_name = osp.basename(video_path)
         frame_list = sorted(glob.glob(osp.join(video_path, '*.png')))
         output_path = osp.join(dst_root, '{}.y4m'.format(video_name))
+
         writer = (ffmpeg
                   .input('pipe:', format='rawvideo', pix_fmt='yuv420p', s='{}x{}'.format(w, h), r=fps)
                   .output(output_path)
                   .overwrite_output()
                   .run_async(pipe_stdin=True)
                   )
+
         for frame_path in frame_list:
             frame_name = osp.basename(frame_path)
             data = cv2.imread(frame_path, cv2.IMREAD_UNCHANGED)
-            # yuv_io.save_yuv(frame, osp.join(dst_root, '{}.yuv'.format(video_name)), '420', mode='ab')
+            data = data.astype(np.float)
             Y = data[:, :, 0]
             U = (data[0::2, 0::2, 1] + data[0::2, 1::2, 1] + data[1::2, 0::2, 1] + data[1::2, 1::2, 1]) / 4
             V = (data[0::2, 0::2, 2] + data[0::2, 1::2, 2] + data[1::2, 0::2, 2] + data[1::2, 1::2, 2]) / 4
